gateway/requirements.runtime.txt
pip>=25.0
setuptools>=70.0
wheel>=0.43
PyYAML>=6.0
cryptography>=3.1
pypdf>=6.0
